+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Replace the `child` member with a simple `title` member on `FixedOnScrollAppbar`.
 - Add a `gradientColors` member on `LitFooter` to customize its background gradient.
+- Add a `contrastColor` member on `LitUserIcon` to customize its background gradient.
 
 ## 1.3.0